BUILD-11417 Extract symlink keeper into its own sub-action
Move `ensureCacheMetricsSymlink` out of credential-guard-post into a new
`symlink-keeper/` sub-action and wire it as the LAST main step of the
top-level composite so its post fires FIRST in the LIFO post phase,
ahead of cache-metrics-post. That ordering is what makes the
`.actions/cache-metrics` workspace symlink survive a nested
actions/checkout sequence (clean → reset --hard → checkout --force)
between cache-metrics-main and cache-metrics-post.

Side effect of the split: credential-guard no longer carries the
metrics-symlink concern, so the `cache-backend == 's3'` gate previously
added to `cache-metrics-prep` is dropped. Metrics now opt in via
CI_METRICS_ENABLED on both backends.

`@symlink-keeper-1.0.0` is a branch ref pinned to this commit's SHA;
after merge to master it is replaced by a tag of the same name on the
same commit.

Files:
- symlink-keeper/{action.yml,dist/main/index.js,dist/post/index.js}: new
- src/symlink-keeper-{main,post}.ts: new (post is a verbatim port of the
  keeper logic previously embedded in credential-guard-post)
- __tests__/symlink-keeper.test.ts: 6 tests (2 main, 4 post)
- src/credential-guard-{main,post}.ts, credential-guard/action.yml:
  reverted to pre-keeper baseline
- credential-guard/dist/{main,post}/index.js: rebuilt
- __tests__/credential-guard.test.ts: 6 keeper-related tests dropped
- package.json: build:keeper-main / build:keeper-post added to `build`
- .gitignore, .pre-commit-config.yaml: add symlink-keeper/dist patterns
- action.yml: drop the s3 gate from cache-metrics-prep; remove the
  cache-metrics-action-path parameter from the credential-guard step;
  add the symlink-keeper step at the end of the main phase
