+++ .gitignore
@@ -15,6 +15,8 @@ credential-guard/dist/**/*.d.ts
 credential-guard/dist/**/*.d.ts.map
 cache-metrics/dist/**/*.d.ts
 cache-metrics/dist/**/*.d.ts.map
+symlink-keeper/dist/**/*.d.ts
+symlink-keeper/dist/**/*.d.ts.map
 
 # Test coverage
 coverage/
