+++ package.json
@@ -4,12 +4,14 @@
   "private": true,
   "description": "S3 Cache action with credential isolation",
   "scripts": {
-    "build": "npm run build:setup && npm run build:guard-main && npm run build:guard-post && npm run build:metrics-main && npm run build:metrics-post",
+    "build": "npm run build:setup && npm run build:guard-main && npm run build:guard-post && npm run build:metrics-main && npm run build:metrics-post && npm run build:keeper-main && npm run build:keeper-post",
     "build:setup": "ncc build src/credential-setup.ts -o credential-setup/dist --license licenses.txt",
     "build:guard-main": "ncc build src/credential-guard-main.ts -o credential-guard/dist/main --license licenses.txt",
     "build:guard-post": "ncc build src/credential-guard-post.ts -o credential-guard/dist/post --license licenses.txt",
     "build:metrics-main": "ncc build src/cache-metrics-main.ts -o cache-metrics/dist/main --license licenses.txt",
     "build:metrics-post": "ncc build src/cache-metrics-post.ts -o cache-metrics/dist/post --license licenses.txt",
+    "build:keeper-main": "ncc build src/symlink-keeper-main.ts -o symlink-keeper/dist/main --license licenses.txt",
+    "build:keeper-post": "ncc build src/symlink-keeper-post.ts -o symlink-keeper/dist/post --license licenses.txt",
     "test": "vitest run",
     "test:watch": "vitest"
   },
